update the project

+++ scripts/buildimg.sh
@@ -1,20 +1,27 @@
 #!/bin/sh
 
-TOP=$(pwd)
-. $TOP/.config
+die() {
+	echo "$*" >&2
+	exit 1
+}
 
+[ -s "./.config" ] || die "please run ./config.sh first."
+
+. ./.config
+
+TOOLS_DIR=$(pwd)/tools/bin
+
+#prepare boot.img
 if [ ! -d "$BOARD/rockdev/Image" ]; then
 	mkdir -p $BOARD/rockdev/Image
 fi
 rm -rf $BOARD/rockdev/Image/boot-linux.img
+$TOOLS_DIR/mkbootimg --kernel $KERNEL_SRC/arch/arm/boot/zImage --ramdisk $INITRD_DIR/../initrd.img --second $KERNEL_SRC/$BOOTIMG_TARGET -o $BOARD/rockdev/Image/boot-linux.img
 
-#prepare boot.img
-mkbootimg --kernel $LINUX_SRC/arch/arm/boot/zImage --ramdisk $INITRD_DIR/initrd.img $BOOTIMG_TARGET -o $BOARD/rockdev/Image/boot-linux.img
-
-cat $TOP/parameter/$BOARD-parameter > $BOARD/rockdev/parameter
-cat $TOP/package-file/$BOARD-package-file > $BOARD/rockdev/package-file
-rm -rf $TOP/rockdev
-ln -s $TOP/$BOARD/rockdev $TOP/rockdev
+cat parameter/$BOARD-parameter > $BOARD/rockdev/parameter
+cat package-file/$BOARD-package-file > $BOARD/rockdev/package-file
+rm -rf rockdev
+ln -s $BOARD/rockdev rockdev
 
 #pack update.img
-cd $TOP/$BOARD/rockdev && ../../scripts/mkupdate.sh && cd - > /dev/null
+./scripts/mkupdate.sh
